Merge style from penny and CSS
Rewrite the main.html.ejs template from CSS and import a new custom style
sheet. This allow to change the appearance of CSS' registration pages
to match the style of Penny UI.
Since we can only change the style and html  of the main template but not
its children ( and therefore the class of element which is needed when using
tailwind framework ), we also import a custom main.css to change the style
of some children's component ( such as the button ).